+++ pkg/lib/authenticationflow/declarative/intent_signup_flow_step_verify.go
@@ -53,7 +53,7 @@ func (i *IntentSignupFlowStepVerify) ReactTo(ctx context.Context, deps *authflow
 	targetStepName := step.TargetStep
 
 	// Find the target step from the root.
-	targetStepFlow, err := FindTargetStep(flows.Root, targetStepName)
+	targetStepFlow, err := authflow.FindTargetStep(flows.Root, targetStepName)
 	if err != nil {
 		return nil, err
 	}
